+++ docker-images/prometheus/cmd/prom-wrapper/change.go
@@ -2,11 +2,16 @@ package main
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"strconv"
 	"time"
 
+	"github.com/go-openapi/strfmt"
 	"github.com/inconshreveable/log15"
+	amclient "github.com/prometheus/alertmanager/api/v2/client"
+	"github.com/prometheus/alertmanager/api/v2/client/silence"
+	"github.com/prometheus/alertmanager/api/v2/models"
 	amconfig "github.com/prometheus/alertmanager/config"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/schema"
@@ -14,6 +19,7 @@ import (
 
 type ChangeContext struct {
 	AMConfig *amconfig.Config // refer to https://prometheus.io/docs/alerting/latest/configuration/
+	AMClient *amclient.Alertmanager
 }
 
 // ChangeResult indicates output from a Change
@@ -101,4 +107,91 @@ func changeSMTP(ctx context.Context, log log15.Logger, change ChangeContext, new
 	change.AMConfig.Global.SMTPRequireTLS = !email.SMTP.DisableTLS
 
 	return
+}
+
+// changeSilences syncs Alertmanager silences with silences configured in observability.silenceAlerts
+func changeSilences(ctx context.Context, log log15.Logger, change ChangeContext, newConfig *subscribedSiteConfig) (result ChangeResult) {
+	// convenience function for creating a prefixed problem - this reflects the relevant site configuration fields
+	newProblem := func(err error) {
+		result.Problems = append(result.Problems, conf.NewSiteProblem(fmt.Sprintf("`observability.silenceAlerts`: %v", err)))
+	}
+
+	var (
+		createdBy = "src-prom-wrapper"
+		comment   = "Applied via `observability.silenceAlerts` in site configuration"
+		startTime = strfmt.DateTime(time.Now())
+		// set 10 year expiry (expiry required, but we don't want it to expire)
+		// silences removed from config will be removed from alertmanager
+		endTime = strfmt.DateTime(time.Now().Add(10 * 365 * 24 * time.Hour))
+		// map configured silences to alertmanager silence IDs
+		activeSilences = map[string]string{}
+	)
+
+	for _, s := range newConfig.SilencedAlerts {
+		activeSilences[s] = ""
+	}
+
+	// delete existing silences that should no longer be silenced
+	existingSilences, err := change.AMClient.Silence.GetSilences(&silence.GetSilencesParams{Context: ctx})
+	if err != nil {
+		newProblem(fmt.Errorf("failed to get existing silences: %w", err))
+		return
+	}
+	for _, s := range existingSilences.Payload {
+		if *s.CreatedBy != createdBy || *s.Status.State != "active" {
+			continue
+		}
+
+		// if this silence should not exist, delete
+		silencedAlert := newSilenceFromMatchers(s.Matchers)
+		if _, shouldBeActive := activeSilences[silencedAlert]; shouldBeActive {
+			activeSilences[silencedAlert] = *s.ID
+		} else {
+			uid := strfmt.UUID(*s.ID)
+			if _, err := change.AMClient.Silence.DeleteSilence(&silence.DeleteSilenceParams{
+				Context:   ctx,
+				SilenceID: uid,
+			}); err != nil {
+				newProblem(fmt.Errorf("failed to delete existing silence %q: %w", *s.ID, err))
+				return
+			}
+		}
+	}
+	log.Info("updating alert silences", "silences", activeSilences)
+
+	// create or update silences
+	for alert, existingSilence := range activeSilences {
+		s := models.Silence{
+			CreatedBy: &createdBy,
+			Comment:   &comment,
+			StartsAt:  &startTime,
+			EndsAt:    &endTime,
+			Matchers:  newMatchersFromSilence(alert),
+		}
+		var err error
+		if existingSilence != "" {
+			_, err = change.AMClient.Silence.PostSilences(&silence.PostSilencesParams{
+				Context: ctx,
+				Silence: &models.PostableSilence{
+					ID:      existingSilence,
+					Silence: s,
+				},
+			})
+		} else {
+			_, err = change.AMClient.Silence.PostSilences(&silence.PostSilencesParams{
+				Context: ctx,
+				Silence: &models.PostableSilence{
+					Silence: s,
+				},
+			})
+		}
+		if err != nil {
+			silenceData, _ := json.Marshal(s)
+			log.Error("failed to update silence", "error", err, "silence", string(silenceData), "existingSilence", existingSilence)
+			newProblem(fmt.Errorf("failed to update silence: %w", err))
+			return
+		}
+	}
+
+	return result
 }